finish CSS, change the format
I finished up CSS on each of the pages. I changed the format on a few parts. I removed the cartoon odin pic. It didn't allow the food picture to be the center of the webpage like I wanted.
In the future, I want to learn how to use CSS better so I can move pictures around on the webpage better. I think that is coming up in the next chapter. For now I am happy with this finished project.